+++ tests/testthat/test-reverse_geocode.R
@@ -20,7 +20,7 @@ test_that("reverse_geocode works", {
 
   # Test with API response mock: sf
   with_mock(
-    "hereR:::.get_content" = function(url) {hereR:::mock$reverse_geocode},
+    "hereR:::.get_content" = function(url) {hereR:::mock$reverse_geocode_response},
     reverse <- reverse_geocode(poi = poi,  results = 3, sf = TRUE),
 
     # Tests